move dbus_adapters package
was: wakepy.io.dbus
now: wakepy.dbus_adapters

remove the old wakepy.io package, as there is no other use for it
currently than the dbus_adapters.
