+++ wakepy/dbus_adapters/__init__.py
@@ -1,0 +1,6 @@
+"""This package contains dbus adapters to be used with wakepy.
+
+Each dbus adapter represents one python dbus library. A dbus adapter should
+subclass the wakepy.core.DbusAdapter, and it may be used as an argument to
+wakepy modes (keep.running & keep.presenting).
+"""
