--- wakepy/io/__init__.py
@@ -1,7 +1,0 @@
-"""This sub-package contains most[1] of the parts of wakepy which do I/O;
-
-* Implementation of dbus adapters
-
-[1] There are some environment variable checks and reading of the platform
-version in the core. 
-"""
